+++ data/migrations/20201104145855_create_clients_table.js
@@ -1,6 +1,6 @@
 exports.up = async function (knex) {
     return await knex.schema.createTable('clients', function (table) {
-        table.increments();
+        table.bigIncrements();
         table.string('name');
         table.string('email').unique();
         table.string('street');
@@ -8,7 +8,7 @@ exports.up = async function (knex) {
         table.string('zip');
         table.string('phone');
         table
-            .integer('user_id')
+            .bigInteger('user_id')
             .unsigned()
             .notNullable()
             .references('id')
